python .\analysis.py
Use a for loop to count the total number of letters, characters, numbers in the text file
The command python .\analysis.py works now.
No need for file input when you run the program analysis.py.
Previously you had to run the command with file input as follows : python .\analysis.py .\iris.csv
This still works but it's not necessary anymore to add .\iris.csv after python .\analysis.py.
But you can if you so wish. Up to you.